// src/components/Organisms/AssetTile/AssetTile.interface.ts
export enum AssetTileLayout {
  auto = 'auto',
  listItem = 'listItem',
  gridItem = 'gridItem',
  compact = 'compact',
}
// ...
export enum AssetStrategy {
  structured = '1',
  timebox = '2',
  none = 'none',
}
// ...
export type MetaItem = {
  numberOfAssets?: string | undefined
  completedStatus?: string | undefined
  duration?: string
  type?: AssetTypeName
}
// ...
export type AssetTypeName =
  | 'Video'
  | 'Article'
  | 'Podcast'
  | 'Online course'
  | 'Playlist'
  | 'Live'
  | 'Sponsored'